refactor: move PhoneNormalizingFormatter to keypad/utils (WT-1026)
Extract formatter into its own file under features/keypad/utils/ with a
barrel export, following the existing package structure convention.
