+++ lib/features/keypad/view/keypad_view.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:webtrit_phone_number/webtrit_phone_number.dart';
 
 import 'package:webtrit_phone/features/call/call.dart';
 import 'package:webtrit_phone/features/call_routing/call_routing.dart';
@@ -10,6 +8,7 @@ import 'package:webtrit_phone/features/call_routing/call_routing.dart';
 import 'package:webtrit_phone/theme/theme.dart';
 
 import '../cubit/keypad_cubit.dart';
+import '../utils/utils.dart';
 import '../widgets/widgets.dart';
 import 'keypad_screen_style.dart';
 
@@ -81,7 +80,7 @@ class KeypadViewState extends State<KeypadView> {
                     showCursor: inputField?.showCursor ?? true,
                     keyboardType: inputField?.keyboardType ?? TextInputType.none,
                     cursorColor: inputField?.cursorColor,
-                    inputFormatters: [_PhoneNormalizingFormatter()],
+                    inputFormatters: [PhoneNormalizingFormatter()],
                   ),
                   BlocBuilder<KeypadCubit, KeypadState>(
                     builder: (context, state) => Text(
@@ -210,22 +209,5 @@ class KeypadViewState extends State<KeypadView> {
     _keypadTextFieldEditableTextState?.userUpdateTextEditingValue(value, SelectionChangedCause.keyboard);
 
     _keypadTextFieldEditableTextState?.hideToolbar();
-  }
-}
-
-class _PhoneNormalizingFormatter extends TextInputFormatter {
-  @override
-  TextEditingValue formatEditUpdate(TextEditingValue oldValue, TextEditingValue newValue) {
-    final normalized = PhoneParser.normalize(newValue.text);
-    if (normalized == newValue.text) return newValue;
-
-    final prefix = newValue.text.substring(0, newValue.selection.baseOffset.clamp(0, newValue.text.length));
-    final normalizedPrefix = PhoneParser.normalize(prefix);
-    final newOffset = normalizedPrefix.length.clamp(0, normalized.length);
-
-    return TextEditingValue(
-      text: normalized,
-      selection: TextSelection.collapsed(offset: newOffset),
-    );
   }
 }
